# openAPI-v1.0.0.yaml
...
openapi: 3.1.0
info:
  title: Inventory Management System - OpenAPI 3.1
  description: |-
    This is an Inventory Management System API documentation. It provides the endpoints for managing items, customers, orders, supplies, and suppliers.
    Some useful links:
    - [The github repository](https://github.com/0xJ1NX/inventory-management-system)
  contact:
    email: omarqalalweh@gmail.com
  version: 1.0.0
servers:
  - url: https://localhost:8080/api/v1
    description: Local server
...
paths:
...
  /customers:
    get:
      tags:
        - customers
      summary: Get all customers
      description: return all customers in json object
      operationId: getCustomers
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  customers:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: int64
                        name:
                          type: string
                        email:
                          type: string
                        phone:
                          type: string
                      required:
                        - id
                        - name
                        - email
                        - phone
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized Access
        "403":
          description: Forbidden Access
        "429":
          description: Too many requests
        "500":
          description: Internal Server Error
    post:
      tags:
        - customers
      summary: Add a new customer
      description: Add a new customer to the inventory
      operationId: addCustomer
      requestBody:
        description: Customer object that needs to be added to the inventory without the id
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
              required:
                - name
                - email
                - phone
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                type: object
                properties:
                  customer:
                    type: object
                    properties:
                      id:
                        type: integer
                        format: int64
                      name:
                        type: string
                      email:
                        type: string
                      phone:
                        type: string
                    required:
                      - id
                      - name
                      - email
                      - phone
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized Access
        "403":
          description: Forbidden Access
        "415":
          description: Unsupported Media Type
        "429":
          description: Too many requests
        "500":
          description: Internal Server Error
  /customers/{id}:
    get:
      tags:
        - customers
      summary: Get a customer by id
      description: Returns a single customer in json object
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          description: ID of the customer to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  customer:
                    type: object
                    properties:
                      id:
                        type: integer
                        format: int64
                      name:
                        type: string
                      email:
                        type: string
                      phone:
                        type: string
                    required:
                      - id
                      - name
                      - email
                      - phone
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized Access
        "403":
          description: Forbidden Access
        "404":
          description: Customer not found
        "429":
          description: Too many requests
        "500":
          description: Internal Server Error
    put:
      tags:
        - customers
      summary: Update a customer
      description: Update a customer in the inventory
      operationId: updateCustomer
      parameters:
        - name: id
          in: path
          description: ID of the customer to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Customer object that needs to be updated in the inventory
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
              required:
                - name
                - email
                - phone
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  customer:
                    type: object
                    properties:
                      id:
                        type: integer
                        format: int64
                      name:
                        type: string
                      email:
                        type: string
                      phone:
                        type: string
                    required:
                      - id
                      - name
                      - email
                      - phone
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized Access
        "403":
          description: Forbidden Access
        "404":
          description: Customer not found
        "415":
          description: Unsupported Media Type
        "429":
          description: Too many requests
        "500":
          description: Internal Server Error
    delete:
      tags:
        - customers
      summary: Delete a customer
      description: Delete a customer from the inventory
      operationId: deleteCustomer
      parameters:
        - name: id
          in: path
          description: ID of the customer to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Customer deleted
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized Access
        "403":
          description: Forbidden Access
        "404":
          description: Customer not found
        "429":
          description: Too many requests
        "500":
          description: Internal Server Error
    patch:
      tags:
        - customers
      summary: Update a customer
      description: Update a customer in the inventory
      operationId: partialUpdateCustomer
      parameters:
        - name: id
          in: path
          description: ID of the customer to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Customer object that needs to be updated in the inventory
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  customer:
                    type: object
                    properties:
                      id:
                        type: integer
                        format: int64
                      name:
                        type: string
                      email:
                        type: string
                      phone:
                        type: string
                    required:
                      - id
                      - name
                      - email
                      - phone
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties: ""
                message:
                  type: string
        "401":
          description: Unauthorized Access
        "403":
          description: Forbidden Access
        "404":
          description: Customer not found
        "415":
          description: Unsupported Media Type
        "429":
          description: Too many requests
        "500":
          description: Internal Server Error
...